samples/python/pubsub/requirements.txt
azure-core==1.13.0
azure-messaging-webpubsubservice==1.0.0b1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
idna==2.10
isodate==0.6.0
msrest==0.6.21
oauthlib==3.1.0
pycparser==2.20
PyJWT==2.0.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.26.4